requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
numpy==2.1.3
pillow==11.3.0
tensorflow-cpu==2.19.1
h5py==3.14.0
pydantic==2.11.7
